# std: Make the CRC32 calculation slightly faster