@Rocknest 2020-09-13
What if `p` is not aligned?