@LemonBoy 2020-09-13
`readIntLittle` shifts down the alignment to 1