@Rocknest 2020-09-13
Oh okay, but isn't that the same as reading byte by byte? Only uses less instructions on x86