@LemonBoy 2020-09-13
You can observe the same speed up by "unrolling" the `readIntLittle` and then xor the resulting u32. I guess (cbf to read the disassembly) the improvement comes from the reduced number of xors.